Implemented runner with averaging and csv output

+++ algorithms/bubble_sort.c
@@ -4,7 +4,7 @@
 
 #include "bubble_sort.h"
 
-void BubbleSort(int* data, const unsigned short int length) {
+void BubbleSort(int* data, const unsigned int length) {
     for(int pass = length - 1; pass >= 0; pass--) {
         for(int i=0; i <= pass - 1 ; i++) {
             if(data[i] > data[i+1]) {
@@ -16,7 +16,7 @@ void BubbleSort(int* data, const unsigned short int length) {
     }
 }
 
-void BubbleSort_QuickExit(int* data, const unsigned short int length) {
+void BubbleSort_QuickExit(int* data, const unsigned int length) {
     int swapped = 1;
     for(int pass = length - 1; pass >= 0 && swapped; pass--) {
         swapped = 0;
