+++ algorithms/bubble_sort.h
@@ -5,7 +5,7 @@
 #ifndef BUBBLE_SORT_H
 #define BUBBLE_SORT_H
 
-void BubbleSort(int* data, const unsigned short int length);
-void BubbleSort_QuickExit(int* data, const unsigned short int length);
+void BubbleSort(int* data, const unsigned int length);
+void BubbleSort_QuickExit(int* data, const unsigned int length);
 
 #endif //BUBBLE_SORT_H
